Refactor with if...else to handle alt configurations
If code is being executed inside the browser and MetaMask is
available then, as before, hijack MetaMask's current provider and use
it to create our own instance of Web3.
Else, set up our own provider that connects to the Rinkeby Test network
through Infura, and use that as the provider for our Web3 instance.